some linting, and SERIALIZABLE_REGISTRY
